+++ Kernel/Bus/USB/xHCI/DeviceTreexHCIController.h
@@ -8,6 +8,7 @@
 
 #include <Kernel/Bus/USB/xHCI/xHCIController.h>
 #include <Kernel/Firmware/DeviceTree/Device.h>
+#include <Kernel/Interrupts/IRQHandler.h>
 
 namespace Kernel::USB::xHCI {
 
@@ -30,6 +31,20 @@ private:
     StringView m_node_name;
     InterruptNumber m_interrupt_number { 0 };
     bool m_using_message_signalled_interrupts { false };
+};
+
+class xHCIDeviceTreeInterrupter final
+    : public xHCIInterrupter
+    , public IRQHandler {
+public:
+    static ErrorOr<NonnullOwnPtr<xHCIDeviceTreeInterrupter>> create(DeviceTreexHCIController&, InterruptNumber irq, u16 interrupter_id);
+
+    virtual StringView purpose() const override { return "xHCI Interrupter"sv; }
+
+private:
+    xHCIDeviceTreeInterrupter(DeviceTreexHCIController& controller, u16 interrupter_id, InterruptNumber irq);
+
+    virtual bool handle_irq() override;
 };
 
 }
