+++ Kernel/Bus/USB/xHCI/PCIxHCIController.h
@@ -10,6 +10,7 @@
 #include <Kernel/Bus/PCI/Definitions.h>
 #include <Kernel/Bus/PCI/Device.h>
 #include <Kernel/Bus/USB/xHCI/xHCIController.h>
+#include <Kernel/Interrupts/PCIIRQHandler.h>
 
 namespace Kernel::USB::xHCI {
 
@@ -43,6 +44,20 @@ private:
     static constexpr PCI::RegisterOffset intel_xhci_usb3_port_routing_mask_offset = static_cast<PCI::RegisterOffset>(0xDC);
 
     bool m_using_message_signalled_interrupts { false };
+};
+
+class xHCIPCIInterrupter final
+    : public xHCIInterrupter
+    , public PCI::IRQHandler {
+public:
+    static ErrorOr<NonnullOwnPtr<xHCIPCIInterrupter>> create(PCIxHCIController&, u16 interrupter_id);
+
+    virtual StringView purpose() const override { return "xHCI Interrupter"sv; }
+
+private:
+    xHCIPCIInterrupter(PCIxHCIController& controller, u16 interrupter_id, InterruptNumber irq);
+
+    virtual bool handle_irq() override;
 };
 
 }
